[LCS] remove test dependency to libra-crypto
Closes: #6709
