+++ common/lcs/Cargo.toml
@@ -16,14 +16,8 @@ serde = { version = "1.0.117", features = ["derive"] }
 
 [dev-dependencies]
 criterion = "0.3.3"
-libra-crypto = { path = "../../crypto/crypto", version = "0.1.0", features = ["fuzzing"] }
-libra-crypto-derive = { path = "../../crypto/crypto-derive", version = "0.1.0" }
 proptest = "0.10.1"
 proptest-derive = "0.2.0"
-rand = "0.7.3"
-
-[features]
-fuzzing = ["libra-crypto/fuzzing"]
 
 [[bench]]
 name = "lcs_bench"
