+++ crypto/crypto/src/unit_tests/mod.rs
@@ -7,5 +7,6 @@ mod cryptohasher;
 mod ed25519_test;
 mod hash_test;
 mod hkdf_test;
+mod lcs_test;
 mod multi_ed25519_test;
 mod noise_test;
